% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibble.R
\name{as_dibble}
\alias{as_dibble}
\alias{as_dibble.data.frame}
\alias{as_dibble.dibble}
\title{Coerce an object to a dibble}
\usage{
as_dibble(x, ...)

\method{as_dibble}{data.frame}(x, dim_names, ...)

\method{as_dibble}{dibble}(x, dim_names, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused, for extensibility.}

\item{dim_names}{A list of dimension names.}
}
\value{
A dibble.
}
\description{
\code{as_dibble()} turns an object into a dimensional data frame called a dibble.
}
