% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibble_measure.R
\name{as_dibble_measure}
\alias{as_dibble_measure}
\alias{as_dibble_measure.dibble_measure}
\alias{as_dibble_measure.dibble}
\alias{as_dibble_measure.grouped_dibble}
\alias{as_dibble_measure.array}
\alias{as_dibble_measure.table}
\title{Coerce an object to a dibble measure}
\usage{
as_dibble_measure(x, ...)

\method{as_dibble_measure}{dibble_measure}(x, ...)

\method{as_dibble_measure}{dibble}(x, ...)

\method{as_dibble_measure}{grouped_dibble}(x, ...)

\method{as_dibble_measure}{array}(x, ...)

\method{as_dibble_measure}{table}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused, for extensibility.}

\item{dim_names}{A list of dimension names.}
}
\value{
A measure for a dibble.
}
\description{
\code{as_dibble_measure()} turns an object into a measure for a dibble.
}
