% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibble_metric.R
\name{as_dibble_metric}
\alias{as_dibble_metric}
\alias{as_dibble_metric.default}
\alias{as_dibble_metric.array}
\alias{as_dibble_metric.dibble_metric}
\title{Coerce an object to a dibble metric}
\usage{
as_dibble_metric(x, ...)

\method{as_dibble_metric}{default}(x, dim_names, ...)

\method{as_dibble_metric}{array}(x, dim_names = NULL, ...)

\method{as_dibble_metric}{dibble_metric}(x, dim_names = NULL, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Unused, for extensibility.}

\item{dim_names}{A list of dimension names.}
}
\value{
A metric for a dibble.
}
\description{
\code{as_dibble_metric()} turns an object into a metric for a dibble.
}
