% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{basic-matrices-arrays}
\alias{basic-matrices-arrays}
\alias{eye}
\alias{eye.default}
\alias{eye.matrix}
\alias{eye.ddf_col}
\alias{eye.tbl_ddf}
\alias{ones}
\alias{ones.default}
\alias{ones.array}
\alias{ones.ddf_col}
\alias{ones.tbl_ddf}
\alias{zeros}
\alias{zeros.default}
\alias{zeros.array}
\alias{zeros.ddf_col}
\alias{zeros.tbl_ddf}
\title{Basic matrices and arrays}
\usage{
eye(x, ...)

\method{eye}{default}(x, y = x, ...)

\method{eye}{matrix}(x, ...)

\method{eye}{ddf_col}(x, ...)

\method{eye}{tbl_ddf}(x, ...)

ones(x, ...)

\method{ones}{default}(x, y = x, ...)

\method{ones}{array}(x, ...)

\method{ones}{ddf_col}(x, ...)

\method{ones}{tbl_ddf}(x, ...)

zeros(x, ...)

\method{zeros}{default}(x, y = x, ...)

\method{zeros}{array}(x, ...)

\method{zeros}{ddf_col}(x, ...)

\method{zeros}{tbl_ddf}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed on to methods.}

\item{y}{A scalar integer.}
}
\value{
A dibble if x is a dibble. Otherwise, returns a matrix or an array.
}
\description{
Create basic matrices and arrays.
}
\details{
These functions override base functions to make them generic. The default
methods call the base versions.
}
