% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibble-package.R
\docType{package}
\name{dibble-package}
\alias{dibble-package}
\title{dibble: Dimensional Data Frames}
\description{
Provides a 'dibble' that implements data cubes (derived from 'dimensional tibble'), and allows broadcasting by dimensional names.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/UchidaMizuki/dibble}
  \item \url{https://uchidamizuki.github.io/dibble/}
  \item Report bugs at \url{https://github.com/UchidaMizuki/dibble/issues}
}

}
\author{
\strong{Maintainer}: Mizuki Uchida \email{uchidamizuki@vivaldi.net}

}
\keyword{internal}
