% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibble.R
\name{dibble}
\alias{dibble}
\title{Build a dimensional data frame}
\usage{
dibble(..., .dim_names = NULL)
}
\arguments{
\item{...}{A set of name-measure pairs.}

\item{.dim_names}{A list of dimension names.}
}
\value{
A dibble.
}
\description{
\code{dibble()} constructs a dimensional data frame called a dibble.
}
\details{
Manipulation functions:
\itemize{
\item mutate()
\item rename()
\item select() & relocate()
\item slice()
}
}
