% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibble_measure.R
\name{dibble_measure}
\alias{dibble_measure}
\title{Build a dibble measure}
\usage{
dibble_measure(x, dim_names = NULL)
}
\arguments{
\item{x}{An object.}

\item{dim_names}{A list of dimension names.}
}
\value{
A measure for a dibble.
}
\description{
\code{dibble_measure()} constructs a measure for a dibble.
}
