% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibble_metric.R
\name{dibble_metric}
\alias{dibble_metric}
\title{Build a dibble metric}
\usage{
dibble_metric(x, dim_names = NULL)
}
\arguments{
\item{x}{An object.}

\item{dim_names}{A list of dimension names.}
}
\value{
A metric for a dibble.
}
\description{
\code{dibble_metric()} constructs a metric for a dibble.
}
