% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pack.R
\name{dibble_pack}
\alias{dibble_pack}
\alias{dibble_unpack}
\title{Pack and unpack dibble}
\usage{
dibble_pack(.data, ..., .names_sep = NULL)

dibble_unpack(data, cols, ..., names_sep = NULL)
}
\arguments{
\item{...}{Passed to \code{tidyr::pack()} or \code{tidyr::unpack()}.}

\item{data, .data}{A dibble}

\item{cols}{Passed to \code{tidyr::unpack()}.}

\item{names_sep, .names_sep}{Passed to \code{tidyr::pack()} or \code{tidyr::unpack()}.}
}
\description{
Pack and unpack dibble
}
