% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extremes.R
\name{extremes}
\alias{extremes}
\alias{pmax}
\alias{pmax.default}
\alias{pmax.ddf_col}
\alias{pmax.tbl_ddf}
\alias{pmin}
\alias{pmin.default}
\alias{pmin.ddf_col}
\alias{pmin.tbl_ddf}
\title{Maxima and Minima}
\usage{
pmax(..., na.rm = FALSE)

\method{pmax}{default}(..., na.rm = FALSE)

\method{pmax}{ddf_col}(..., na.rm = FALSE)

\method{pmax}{tbl_ddf}(..., na.rm = FALSE)

pmin(..., na.rm = FALSE)

\method{pmin}{default}(..., na.rm = FALSE)

\method{pmin}{ddf_col}(..., na.rm = FALSE)

\method{pmin}{tbl_ddf}(..., na.rm = FALSE)
}
\arguments{
\item{...}{Dibbles, numeric or character arguments.}

\item{na.rm}{a logical indicating whether missing values should be removed.}
}
\value{
A dibble if \code{...} are dibbles. See \code{\link[base:Extremes]{base::pmax()}} and \code{\link[base:Extremes]{base::pmin()}}
for the return value of the default method.
}
\description{
Returns the parallel maxima and minima of the input values.
}
\details{
These functions override base functions to make them generic. The default
methods call the base versions.
}
\seealso{
\code{\link[base:Extremes]{base::pmax()}}, \code{\link[base:Extremes]{base::pmin()}}.
}
