% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matrix.R
\name{\%*\%}
\alias{\%*\%}
\title{Matrix Multiplication}
\usage{
x \%*\% y
}
\arguments{
\item{x}{Numeric or complex dibble, matrices or vectors.}

\item{y}{Numeric or complex dibble, matrices or vectors.}
}
\value{
A dibble if x is a dibble. See \code{\link[base:matmult]{base::\%*\%}} for the return
value of the default method.
}
\description{
Multiplies two matrices, if they are conformable.
}
\details{
\code{\%*\%} overrides \code{\link[base:matmult]{base::\%*\%}} to make it generic. The default method
calls the base version.
}
\seealso{
\code{\link[base:matmult]{base::\%*\%}}
}
