% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibble_measure.R
\name{is_dibble_measure}
\alias{is_dibble_measure}
\title{Test if the object is a dibble measure}
\usage{
is_dibble_measure(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical.
}
\description{
Test if the object is a dibble measure
}
