% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibble_metric.R
\name{is_dibble_metric}
\alias{is_dibble_metric}
\title{Test if the object is a dibble metric}
\usage{
is_dibble_metric(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical.
}
\description{
Test if the object is a dibble metric
}
