% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_ddf.R
\name{is_grouped_ddf}
\alias{is_grouped_ddf}
\title{Test if the object is a grouped dibble}
\usage{
is_grouped_ddf(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical.
}
\description{
Test if the object is a grouped dibble
}
