% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grouped_dibble.R
\name{is_grouped_dibble}
\alias{is_grouped_dibble}
\title{Test if the object is a grouped dibble}
\usage{
is_grouped_dibble(x)
}
\arguments{
\item{x}{An object.}
}
\value{
A logical.
}
\description{
Test if the object is a grouped dibble
}
