% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{tidyr-data}
\alias{tidyr-data}
\alias{population_dibble}
\alias{table1_dibble}
\alias{us_rent_income_dibble}
\alias{fish_encounters_dibble}
\title{Example dibbles from tidyr example data}
\source{
\url{https://www.who.int/teams/global-tuberculosis-programme/data}

\url{https://fishsciences.github.io/post/visualizing-fish-encounter-histories/}
}
\usage{
population_dibble

table1_dibble

us_rent_income_dibble

fish_encounters_dibble
}
\description{
Example dibbles transformed from example data in tidyr
(\code{\link[tidyr:who]{tidyr::population}}, \code{\link[tidyr:table1]{tidyr::table1}}, \code{\link[tidyr:us_rent_income]{tidyr::us_rent_income}}, and \code{\link[tidyr:fish_encounters]{tidyr::fish_encounters}}).
}
\keyword{datasets}
