% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibble.R
\name{as_dibble}
\alias{as_dibble}
\alias{as_dibble.data.frame}
\title{Coerce a data frame to a dibble}
\usage{
as_dibble(x, ...)

\method{as_dibble}{data.frame}(x, dim_names, ...)
}
\arguments{
\item{x}{A data frame or a dibble.}

\item{...}{Unused, for extensibility.}

\item{dim_names}{A list of dimension names.}
}
\value{
A dibble.
}
\description{
\code{as_dibble()} turns a data frame into a dimensional data frame called a dibble.
}
