% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibble.R
\name{dibble_by}
\alias{dibble_by}
\title{Constructs a dibble by one or more variables}
\usage{
dibble_by(x, ...)
}
\arguments{
\item{x}{A data frame or a dibble.}

\item{...}{Variables.}
}
\description{
\code{dibble_by()} constructs a dibble by one or more variables.
}
