% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dibble.R
\name{nrow-ncol}
\alias{nrow-ncol}
\alias{nrow}
\alias{nrow.default}
\alias{nrow.dibble}
\alias{ncol}
\alias{ncol.default}
\alias{ncol.dibble}
\title{The number of rows/columns}
\usage{
nrow(x, ...)

\method{nrow}{default}(x, ...)

\method{nrow}{dibble}(x)

ncol(x)

\method{ncol}{default}(x)

\method{ncol}{dibble}(x)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed on to methods.}
}
\value{
An integer or \code{NULL}.
}
\description{
\code{nrow()} and \code{ncol()} return the number of rows or columns present in x.
}
