% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_ddf.R
\name{nrow-ncol}
\alias{nrow-ncol}
\alias{nrow}
\alias{nrow.default}
\alias{nrow.tbl_ddf}
\alias{ncol}
\alias{ncol.default}
\alias{ncol.tbl_ddf}
\title{The number of rows/columns}
\usage{
nrow(x, ...)

\method{nrow}{default}(x, ...)

\method{nrow}{tbl_ddf}(x, ...)

ncol(x, ...)

\method{ncol}{default}(x, ...)

\method{ncol}{tbl_ddf}(x, ...)
}
\arguments{
\item{x}{An object.}

\item{...}{Other arguments passed on to methods.}
}
\value{
An integer or \code{NULL}.
}
\description{
\code{nrow()} and \code{ncol()} return the number of rows or columns present in x.
}
