% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl_ddf.R
\name{row-colnames}
\alias{row-colnames}
\alias{rownames}
\alias{rownames.default}
\alias{rownames.tbl_ddf}
\alias{colnames}
\alias{colnames.default}
\alias{colnames.tbl_ddf}
\title{Row and column names}
\usage{
rownames(x, ...)

\method{rownames}{default}(x, ...)

\method{rownames}{tbl_ddf}(x, ...)

colnames(x, ...)

\method{colnames}{default}(x, ...)

\method{colnames}{tbl_ddf}(x, ...)
}
\arguments{
\item{x}{A matrix-like object.}

\item{...}{Other arguments passed on to methods.}
}
\value{
A list of row/column names.
}
\description{
Retrieve or set the row or column names of a matrix-like object.
}
