% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row-col.R
\name{row-colnames}
\alias{row-colnames}
\alias{rownames}
\alias{rownames.default}
\alias{rownames.ddf_col}
\alias{rownames.tbl_ddf}
\alias{colnames}
\alias{colnames.default}
\alias{colnames.ddf_col}
\alias{colnames.tbl_ddf}
\title{Row and column names}
\usage{
rownames(x, ...)

\method{rownames}{default}(x, ...)

\method{rownames}{ddf_col}(x, ...)

\method{rownames}{tbl_ddf}(x, ...)

colnames(x, ...)

\method{colnames}{default}(x, ...)

\method{colnames}{ddf_col}(x, ...)

\method{colnames}{tbl_ddf}(x, ...)
}
\arguments{
\item{x}{A matrix-like object.}

\item{...}{Other arguments passed on to methods.}
}
\value{
A list of row/column names.
}
\description{
Retrieve or set the row or column names of a matrix-like object.
}
\details{
These functions override base functions to make them generic. The default
methods call the base versions.
}
